+++ contracts/multisig-prover/src/encoding/abi.rs
@@ -8,7 +8,7 @@ use k256::{elliptic_curve::sec1::ToEncodedPoint, PublicKey};
 use sha3::{Digest, Keccak256};
 
 use axelar_wasm_std::operators::Operators;
-use multisig::{key::Signature, msg::Signer, workerset::WorkerSet};
+use multisig::{key::Signature, msg::Signer, worker_set::WorkerSet};
 
 use crate::{
     error::ContractError,
