+++ contracts/multisig-prover/src/encoding/mod.rs
@@ -7,7 +7,7 @@ use cosmwasm_std::{HexBinary, Uint256};
 use sha3::{Digest, Keccak256};
 
 use connection_router::state::Message;
-use multisig::{key::Signature, msg::Signer, workerset::WorkerSet};
+use multisig::{key::Signature, msg::Signer, worker_set::WorkerSet};
 
 use crate::{
     error::ContractError,
