+++ contracts/multisig/src/lib.rs
@@ -6,7 +6,7 @@ pub mod msg;
 pub mod signing;
 pub mod state;
 pub mod types;
-pub mod workerset;
+pub mod worker_set;
 
 #[cfg(feature = "secp256k1")]
 mod secp256k1;
